+++ Integrations/EDL/EDL.py
@@ -51,7 +51,7 @@ def get_params_port(params: dict = demisto.params()) -> int:
 
 def refresh_edl_context(indicator_query: str, limit: int = 0) -> str:
     """
-    Refresh the cache values and format using an indicator_query to call demisto.findIndicators
+    Refresh the cache values and format using an indicator_query to call demisto.searchIndicators
     Returns: List(IoCs in output format)
     """
     now = datetime.now()
@@ -69,7 +69,7 @@ def save_context(now: datetime, out_dict: dict):
 
 def find_indicators_to_limit(indicator_query: str, limit: int) -> list:
     """
-    Finds indicators using demisto.findIndicators
+    Finds indicators using demisto.searchIndicators
     """
     iocs, _ = find_indicators_to_limit_loop(indicator_query, limit)
     return iocs[:limit]
@@ -78,13 +78,13 @@ def find_indicators_to_limit(indicator_query: str, limit: int) -> list:
 def find_indicators_to_limit_loop(indicator_query: str, limit: int, total_fetched: int = 0, next_page: int = 0,
                                   last_found_len: int = PAGE_SIZE):
     """
-    Finds indicators using while loop with demisto.findIndicators, and returns result and last page
+    Finds indicators using while loop with demisto.searchIndicators, and returns result and last page
     """
     iocs: List[dict] = []
     if not last_found_len:
         last_found_len = total_fetched
     while last_found_len == PAGE_SIZE and limit and total_fetched < limit:
-        fetched_iocs = demisto.findIndicators(query=indicator_query, page=next_page, size=PAGE_SIZE).get('iocs')
+        fetched_iocs = demisto.searchIndicators(query=indicator_query, page=next_page, size=PAGE_SIZE).get('iocs')
         iocs.extend(fetched_iocs)
         last_found_len = len(fetched_iocs)
         total_fetched += last_found_len